Add authorization to portfolio goal

+++ app/Http/Controllers/Portfolio/PortfolioGoalController.php
@@ -12,6 +12,7 @@ class PortfolioGoalController extends Controller
 {
     public function index(PortfolioUnit $portfolioUnit)
     {
+        $this->authorize('view', $portfolioUnit);
         $links = $portfolioUnit->goals;
         $editRoute = route('portfolios.goals.edit', ['portfolio_unit' => $portfolioUnit->pid, 'link' => 'LLIINNKK']);
         return view('portfolios.goals.index', compact('portfolioUnit', 'links', 'editRoute'));
@@ -19,13 +20,17 @@ class PortfolioGoalController extends Controller
 
     public function create(PortfolioUnit $portfolioUnit)
     {
+        $this->authorize('create', $portfolioUnit);
         $link = new Link;
         $formAction = route('portfolios.goals.store', ['portfolio_unit' => $portfolioUnit->pid]);
-        return view('links.edit', compact('link', 'formAction'));
+        $parentModel = $portfolioUnit;
+        $deleteRoute = '';
+        return view('links.edit', compact('link', 'formAction', 'deleteRoute', 'parentModel'));
     }
 
     public function store(LinkRequest $request, PortfolioUnit $portfolioUnit)
     {
+        $this->authorize('view', $portfolioUnit);
         $link = new Link($request->validated());
         $link->linkable_subtype = Link::SUBTYPE_PORTFOLIO_GOAL;
         $portfolioUnit->links()->save($link);
@@ -34,9 +39,11 @@ class PortfolioGoalController extends Controller
 
     public function edit(PortfolioUnit $portfolioUnit, Link $link)
     {
+        $this->authorize('update', $portfolioUnit);
         $deleteRoute = route('portfolios.goals.destroy', ['portfolio_unit' => $portfolioUnit->pid, 'link' => $link->pid]);
         $formAction = route('portfolios.goals.update', ['portfolio_unit' => $portfolioUnit->pid, 'link' => $link->pid]);
-        return view('links.edit', compact('link', 'deleteRoute', 'formAction'));
+        $parentModel = $portfolioUnit;
+        return view('links.edit', compact('link', 'deleteRoute', 'formAction', 'parentModel'));
     }
 
     public function update(LinkRequest $request, PortfolioUnit $portfolioUnit, Link $link)
@@ -47,6 +54,7 @@ class PortfolioGoalController extends Controller
 
     public function destroy(PortfolioUnit $portfolioUnit, Link $link)
     {
+        $this->authorize('delete', $portfolioUnit);
         $link->delete();
         return Redirect::route('portfolios.goals.index', ['portfolio_unit' => $portfolioUnit->pid]);
     }
